Add lftools main function
The idea of lftools is that it will be the main function for shell
scripts and tools we build. All the tools can be sourced using:

    source $SCRIPT_DIR/<toolname>

If using this method the tools should be created as shell functions so
that they are easily consumable as a library rather than an independent
tool.

Change-Id: I3d3f7c4a548177dcb94ef786a201f308f69a3e0a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>