Update patch set 2
Patch Set 2:

Actually I just realized. If you run coala with the YAMLLintBear in interactive mode. It will provide a recommended solution and if you would like to correct it for you.

I think with that we won't need to maintain a script to fix linting. Would be worth investigating.

Patch-set: 2