Reformat lftools cmd docs for sphinx and click
Sphinx and Click have very different docstring formats unfortunately
which makes it difficult for us to share the same documentation in both
places without breaking the format of the other. This patch proposes a
compromise where we use sphinxcontrib-programoutput to copy the output
of Click --help documentation and place it inside of Sphinx. Not the
best but ensures we only write documentation once.

This patch also fixes the documentation for the Version command as it
was inaccurate or misleading in some places.

Also disable PyDocLint rule for backslashes since Click uses \b as a
keyword for text wrapping.

Change-Id: Ica09079d339065912fb5a78dccb24ee9a3063abc
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>