Update patch set 7
Patch Set 7:

(1 comment)

I think think the suggestion could be a useful feature so I opened another Gerrit to track it as a todo item. I think the code here as it exists should be merged as it's complete.

Also in OpenDaylight we mark officially supported images as "Public" in the private cloud so when running cleanup with --hide-public it does not cleanup the supported images. Only regularly created images that the sandbox produces but might never be used.

Patch-set: 7
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0