Update patch set 35
Patch Set 35:

> > (1 comment)
 > 
 > While using the CLI options through click does not fail, since the
 > positional params are processed and passed at the end.
 > 
 > With the way we have scripted the usage/CLI, its something of a
 > limitation, since we should be processing everything through
 > getopts imho, if we plan on invoking the script directly in jobs,
 > since we can easily be handled in the wrapper in lftools ?

We should think about the shell scripts and the wrapper as 2 independent components. Don't use the wrapper s an excuse to have bugs in the scripts. As it stands today the message is plainly wrong. If someone calls the script and passes more arguments than necessary they will confusingly get a message stating that they have not passed enough arguments.

This is not a good user experience. If we want to not allow more than the required arguments than we should at least return an error message that helps the user troubleshoot. Just because the Python wrapper passes exactly the right number of arguments and won't hit this condition is not an excuse to have a poorly defined error message being returned to the user. More information is good to help troubleshoot when things go wrong.

Patch-set: 35