Update patch set 5
Patch Set 5: Code-Review-1

We should setup a universal logger in a seperate patch that this patch can reuse. Then it makes it so it's a 1 liner to add a logger to very Python module.

We have another Jira that's tracking that work and I've implemented loggers in the past. You can look at my freeseer project on GitHub or spectrometer if you want to try to copy the implementation?

Patch-set: 5
Label: Code-Review=-1