Update patch set 25
Patch Set 25: -Code-Review

Even though this script is not too urgent, I would like to be able to merge this v1.0 that fixes the most common yamllint errors for the user:
-wrong indentation
-line-lenght
-new-line-at-end-of-file
-too many spaces
-braces
-beginning of the file

I would like to add more cases to handle to the script as a side project to help me also add learning value. But for now, I think it would be nice to have this committed with the limited functionality to at least help fixing these cases in 1 shot, specially for repos in which yamllint checks are just being introduced and the existing files need to be scanned for the first time. That is what made me write this in first place when I started scanning hyp :)

Patch-set: 25
Label: Code-Review=0