Update patch set 1
Patch Set 1: Code-Review-1

If we're including a logging.ini we now then have to worry about where this file lives and how it's packaged. Will logging.ini get installed as part of `pip install lftools` and do we have to do something to make the code pick it up?

Patch-set: 1
Label: Code-Review=-1