Update patch set 1
Patch Set 1:

Thanks so much for your comments Thanh!
reading about options on how to setup the logger, I found this option to be a simple one and wanted to bring it up. I was also contemplating using a yaml file but I read in one tutorial that logger.ini was more flexible. (I am still trying to get my learning on python since I am not too fluent). 

If we set this logger.ini as part of ~/.config/lftools/logging.ini, will this be copied over by the shell scripts? or is there a better way to do this directly in the setup.py?

For being able to have the scripts to use it, there is not much installation needed as far as I know from what I was investigating. Only we need to make sure we add "logging.config.fileConfig('<path of logging.ini>')"

Patch-set: 1
Label: Code-Review=0