Update patch set 1
Patch Set 1:

> reading about options on how to setup the logger, I found this
 > option to be a simple one and wanted to bring it up. I was also
 > contemplating using a yaml file but I read in one tutorial that
 > logger.ini was more flexible. (I am still trying to get my learning
 > on python since I am not too fluent).
 > 

Ini is fine since logger module can read the ini.

 > If we set this logger.ini as part of ~/.config/lftools/logging.ini,
 > will this be copied over by the shell scripts? or is there a better
 > way to do this directly in the setup.py?

No, it won't the user will have to create it or you will have to write code to copy it over. But I think implementation wise it should still work even without it.

 > For being able to have the scripts to use it, there is not much
 > installation needed as far as I know from what I was investigating.
 > Only we need to make sure we add "logging.config.fileConfig('<path
 > of logging.ini>')"

We need to write code to check the 3 locations I suggested. Maybe pseudo code will help explain:

 if exists(~/.config/lftools/logging.ini):
     log_file = ~/.config/lftools/loggging.ini
 elif exists(/etc/lftools/logging.ini):
     log_file = /etc/lftools/logging.ini
 else:
     log_file = /default/path/to/site-packages/virtualenv as packaged by lftools
 
 logging.config.fileConfig(log_file)

This way we are checking 3 locations for a log file and if the user did not provide any then default to the one we package with LFTOOLs (pip install lftools should install one to a default location somewhere in site-packageds). You will need to edit MANIFEST.in so that setup.py will know to package the file somewhere. More info on how Python packaging works here:

https://docs.python.org/3/distutils/setupscript.html

Patch-set: 1