Trim trailing slashes from URL inputs
Trailing slashes in URLs causes the staging code to upload to
//org.opendaylight.project instead of /org.opendaylight.project which
fails our URL pattern matching target defined in Nexus since Nexus
thinks // as a blank directory in the path. We need to make sure
extraneous / does not get pushed in by the user so trim it off if they
do.

Issue: RELENG-351
Change-Id: I264fec5c41e89d588e03a9c447917ee8a3623c62
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>