Add lftools command to remove specific server
This command also supports an optional minutes parameter which
can be used to ignore the server if it is not older than a certain
number of minutes.

This is useful for cases like Jenkins orphan server cleanup scripts
so that they do not try to remove servers that have just been
spawned and have not yet connected to Jenkins causing the server to
be mistakenly marked as orphaned.

Issue: RELENG-780
Change-Id: I7f12218e0ac730ff43def4ec9b6133b4cc8ee0c8
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>