Update patch set 9
Patch Set 9:

It WILL break the HEAT jobs in OpenDaylight. So we need to coordinate the release with OpenDaylight. I'm not expecting another lftools release for awhile so we got time to coordinate that. It's safe to merge into master.

Patch-set: 9
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0