Update patch set 34
Patch Set 34: Code-Review-1

(3 comments)

Can we consistently use 4 space indents for the shell scripts so that they are easier to read?

Do the ldap and info file changes need to be merged together? if not maybe we can break them out into 2 patches so that we can merge them faster / have smaller amounts of things to review.

This is fine as is too if it's too hard to do.

Patch-set: 34
Label: Code-Review=-1