Disable fast fail on parallel sigul
Sigul bridges seem to possibly have an issue with a lot of in parallel
signing requests. This is either an issue with the bridge, the server,
or a disconnected across the network VPN backhaul between the server and
bridge. Instead of having the signing just outright fail on the first
problem, we should try a few times.

Change-Id: I98a5c8adeb25e892899a221dc2078a9fc779f74e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>