Update patch set 6
Patch Set 6: Code-Review-1

Forgot to mention we require reno release notes for patches that do a significant amount of work now. Just run this:

 tox -e reno -- new <slug>

Where slug is a short (no spaces) words to describe the change. Eg.

 tox -e reno -- new docker-image-cmds

or something... The edit the RST file that is created to describe your change.

Patch-set: 6
Label: Code-Review=-1