Python script to copy Nexus3 images to Docker Hub
This script will fetch all repositories that matches
an org, as well as a repo pattern (if provided) from Nexus3.
From these repositories, it will check if all valid tags
have been copied to Docker Hub, and if not, copy them
there (and create a possible missing repository at the
same time).

Change-Id: I6fd7658d85d696d39042f8210c7e2098d47f3bbc
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>