Reword some of the upload output
Printing file and file upload URL on the same line makes the output a
little too busy and hard to read. We should print the upload URL once
before the upload and after only print the files because we can assume
the path will be the same from that point on.

This makes the output quite a bit easier to parse.

Issue: RELENG-1512
Change-Id: I3b8a0ac07b0cad72706a080a8abf3fd79e2dffed
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>