Only skip files start with maven-metadata.xml
Ensure we are only skipping files that start with maven-metadata.xml*
Having the word "maven-metadata" in our artifact is fine as long as it
is not a Maven 2 repo file which is exactly 'maven-metadata.xml' the
trialing * is to ensure it does not pick up the checksum files too.

Also reword the function docs to be more clear about what it does.

Issue: RELENG-1512
Change-Id: Icc20e4581a74625506ac9a0ff5d5e17b45034ccc
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>