Integrate Spawn_New_Project.sh into lftools
This change opens a place for

lftools gerrit create [OPTIONS] USER GERRIT_URL REPO

The shell script called currently only checks if gerrit can be
connected to.

Eventually this should create a project and setup permissions.

This is the minimal commit so we can discuss what we should call the
functions etc.

Issue: RELENG-1524

Change-Id: Id5bb3c52918c5fcad9e10b32ee14a90eb0b0d3ff
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>