Update patch set 1
Patch Set 1:

Since we have to support Windows systems now I'd like to see all new contributions to lftools be written in Python unless we have a very good reason not to.

Considering there's not too much to this gerrit_create yet it shouldn't be too much of a pain to rewrite this in Python?

Patch-set: 1