Update patch set 1
Patch Set 1:

I do not want to write this with GitPython and Paramiko

I have no need for this to be supported in windows.
This is a perfect use case for bash.

The commands that need to be run are bash commands.
git cloning, fetching, git config add etc.  

This is only meant to run on the internal jenkins server.
I'm happy to do this in jenkins jobs and not include it in lftools.

Patch-set: 1
Label: Code-Review=0