Python script to copy Nexus3 images to Docker Hub
This script will fetch all repositories that matches
an org, as well as a repo pattern (if provided) from Nexus3.
From these repositories, it will check if all valid tags
have been copied to Docker Hub, and if not, copy them
there (and create a possible missing repository at the
same time).

This is the third part, which contains the Project class

Change-Id: I1b59a2452dd0f415cefe38a2d496f263bdf5d7b8
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>