Python script to copy Nexus3 images to Docker Hub
This script will fetch all repositories that matches
an org, as well as a repo pattern (if provided) from Nexus3.
From these repositories, it will check if all valid tags
have been copied to Docker Hub, and if not, copy them
there (and create a possible missing repository at the
same time).

This is the fourth part, which contains fetching nexus3
catalog

Change-Id: Ie627ad04cc0d84304b585d126ed1e468c95661b7
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>