Python script to copy Nexus3 images to Docker Hub
This script will fetch all repositories that matches
an org, as well as a repo pattern (if provided) from Nexus3.
From these repositories, it will check if all valid tags
have been copied to Docker Hub, and if not, copy them
there (and create a possible missing repository at the
same time).

This is the fifth part, which contains print functions

Change-Id: I23649c998f5a3b1e81edced6538741cbbd51a0d2
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>