Python script to copy Nexus3 images to Docker Hub
This script will fetch all repositories that matches
an org, as well as a repo pattern (if provided) from Nexus3.
From these repositories, it will check if all valid tags
have been copied to Docker Hub, and if not, copy them
there (and create a possible missing repository at the
same time).

This is the seventh, and final part, which contains the cli
part.

Change-Id: I629d575730abd73425dacd5e694a30cadcd85657
Signed-off-by: Bengt Thuree <bthuree@linuxfoundation.org>