Allow requesting serial signing
Sigul-signing directories containing a lot of large artifacts may not be
optimal, because whether serial or parallel, the same amount of
network IO is required and the job takes about the same amount of time
in the end. This allows passing "serial" as a third argument to sign_dir
to specifically request serial operation.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>