add-remove-users matches ldap group to INFO file
I have abandoned the idea of using a diff on the INFO file as a driver for
making changes to ldap. It was messy for new info files, as there was no
previous diff. which meant if a user was removed, we had no way of
knowing. Now the script is idempotent. Whatever is in the INFO.yaml will
be reflected in the LDAP file.

example: you give it an info file and a group and it will
make the nessesary changes:

lftools lfidapi add-remove-committers path/to/INFO.yaml ldap-group

Change-Id: Ie1546f3e80e826742612b9f32f91dc6ce3360d0c
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>