Github audit create and list repos
audit       List users for an Org that do not have 2fa enabled.
create      Create a Github repo within an Organization.
list-org    List owners, teams and their users for an Org.
list-repos  List an Organization's GitHub repos.

Github token is read from
~/.config/lftools/lftools.ini
In the form of:
[github]
token =
This has been added to the doc

ISSUE: RELENG-1526
Change-Id: I558eb11897b555d26fac5abacb660d4bdb630a12
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>