Preliminart jira api tools
create-project
set-owner (ldap group that is owner for a project)
and
get-scheme

get-scheme is needed to find the values for create-project
and set-owner
for example to set owner for a project
in opnfv the Administrators group is 1002
this is probably the default, but I have to check all the
jiras first. at least if your project adds ldap groups as
jira-devlopers you can adapt this to 1001 or whatever

This is implemented with basic auth, base64 encoded in the
lftools config file.
echo -n "user:pass" | base64

I will fix to use oauth, but it means adding jira application
links in all jiras.
Jira settings (cog icon) > Applications > Application links

Change-Id: I6efbf15e3be0d9389a995d7674ea4b18ee2f0f0f
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>