Preliminary jira api tools
Commands:
create-project
set-owner (ldap group that is owner for a project)
get

lftools jira get is needed to find the values for
create-project and set-owner

For example: to set owner for a project
in opnfv the Administrators group is 1002
this is probably the default, but I have to check all the
jiras first. at least if your project adds ldap groups as
jira-devlopers you can adapt this to 1001 or whatever

This is implemented with basic auth, base64 encoded in the
lftools config file [jira] section
echo -n "user:pass" | base64

example:
[jira]
base64 = dXNlcjpwYXNzd29yZGZvbw==

This could be fixed to use oauth, but it means adding jira application
links in all jiras.
Jira settings (cog icon) > Applications > Application links

docs for oauth:
https://developer.atlassian.com/server/jira/platform/oauth/
code for oauth:
https://gist.github.com/Aricg/1e65cc3e4dde3ce23f30ee414a11c469

Change-Id: I6efbf15e3be0d9389a995d7674ea4b18ee2f0f0f
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>