Update patch set 6
Patch Set 6:

this will still work with sso/auth0, it will require a non sso admin account to be added to each jira, but this is essentially the same as adding an app to each jira for oauth, and much simpler. I talked to Trecor regarding ITX and there's nothing in the roadmap for creating jira projects or assigning ldap groups. (only spinning up jiras) Also, the admin panel's main focus it to allow project managers to manage more projects. We RE's will still need lftools and INFO files to facilitate our management.

Patch-set: 6
Label: Code-Review=0