Update patch set 1
Patch Set 1: Code-Review-1

Too many topics in 1 patch. Please split the additional compression change into a separate patch so that we can discuss it separately.

The reason we did not implement xz when we originally set this infrastructure up is that xz files do not automatically open in the web browser. You'll have to download it first and then you can view it.

I think the ability for the files to automatically open in the web browser is very useful. I'm not sure if xz is supported today but it wasn't (or we did not know how to support it) at the time we designed the log archive system.

Patch-set: 1
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1