Update patch set 1
Patch Set 1: Code-Review-1

As I mentioned in the previous commit the reason we did not implement xz when we originally set this infrastructure up is that xz files do not automatically open in the web browser. You'll have to download it first and then you can view it.

The ability for the files to automatically open in the web browser is very useful. I'm not sure if xz is supported today but if we cannot navigate the log server and automatically load the file in the web browser without being forced to download it first then I am opposed to this change.

Anil, can you push an xz file to a log server and show that it can be opened in a web browser?

Patch-set: 1