Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm onboard with dropping support for Python 2 support altogether as well. I think it would be good to bump our CI base support level to distro versions that ship with Py 3.6.

Uhh... that would exclude C7 since Py3.6 is coming in from EPEL and not directly. C8 should be arriving soonish now that RHEL8 is out and it defaults to Py3.6 so I'll have no issues with us bumping our base to C8 as soon as it's out.

Patch-set: 1