Update patch set 1
Patch Set 1:

No we shouldn't put this requirement into lftools. It is not a dependency of lftools and will now cause lftools to pull in numpy on install as an unnecessary requirement. It's being pulled in by the CommitMsgBear linter from Coala in tox.ini. The dependency should be put in tox.ini if we want to keep using the CommitMsgBear. Alternatively if we don't want to use numpy we can switch to pre-commit gitlint instead.

Patch-set: 1