Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'm not sure this patch address the issue at all. There error is not that there are multiple servers listed in a `jenkins_jobs.ini` file, but that a `[jenkins]` section is missing along with values for `url`, `user`, `password`.
> 
> As far as I know, JJB supports not including a `[jenkins]` section (if this is not the case, the code should bail early with a clearer error that it's missing) and the code currently doesn't support this.
> 
> The simplest approach to this I believe is as follows:
> 1) Check for the `[jenkins]` section using has_section[1]
>    If it's not found, add one[2] along with defaults for url, user, password
> 2) If `[jenkins]` is found pull the values, using defaults if any of url, user, or password are not set.[3]
> 
> [1] https://docs.python.org/3/library/configparser.html#configparser.ConfigParser.has_section
> [2] https://docs.python.org/3/library/configparser.html#configparser.ConfigParser.add_section
> [3] https://docs.python.org/3/library/configparser.html#configparser.ConfigParser.get

I disagree, the tool should not go and edit the local 'jenkins.ini' file and add a new '[jenkins]' section, then every time the tool is run it goes and modified the user's jenkins.ini file. 

Instead, its simpler to print an error (with the available server sections in the .ini only in --debug mode) next run the user can select the correct one and _not_ crash, 

We are simply addressing an incorrect way of reading a dict below, which assumes that 'jenkins' key exists in all cases, while in some cases it does not.

63 user = config.get('jenkins', 'user')
64 password = config.get('jenkins', 'password')

Patch-set: 1