Update patch set 5
Patch Set 5: -Code-Review

Sorry I did not realize this was about the --help function until now.

IMO the proposed code does not resolve the underlying issue which is code is being executed when --help is called when it really should just print the help menu and exit.

The solution proposed is bandaid solution which will break down when there's a new piece of code added to __init__ which has dependency on another file or configuration.

I'm not sure if Click has a recommended method to prevent __init__() from being run. Anyway I will leave the call to LF folks on this one as ya'll will be the ones to support this code ;)

Patch-set: 5
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0