Revert plugin listing print change
Change 5df955f24bc0154f7069ecfc188311052e67febc had a minor change that
slipped through the code review changing the Jenkins plugin listing from
long names to using the short names. This however, is not utilized at
all by the security listing function that was the primary portion of
this change.

This modification causes the jenkins configuration management merge job
to fail on updates to the OpenStack cloud configuration.

Change-Id: I04cfbfe5a49904875a734e223bda13c6164df664
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>