Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> (2 comments)

Chris, thanks for catching the tabs. Re the docs, the rst files are there to generate the docs based on the docstrings in the code. lftools/api/endpoints/readthedocs.py has the actual documentation. Would you take a look when you have a moment and tell me if you think it is adequate?

Patch-set: 9
Label: Code-Review=0