Update patch set 1
Patch Set 1:

(8 comments)

Besides needing some cleanup and better error handling, I think there is a logic error. Is it possible that the script calling this script ignores it's status ? I certainly might be wrong, but I think this script will almost always 'exit 1'.

Maybe we need to create a ticket for this.....

Patch-set: 1