Update patch set 12
Patch Set 12: Code-Review-1

I know you're working against a beta edition of the API, but I feel like we shouldn't be implementing the RTD access module this way. It would make sense for that to be it's own module on PyPi for others to be able to use as well!

Did you check to see if there an RTD module available already that we could use / perhaps work with to get the new API features built out in?

I see no unit tests in here for any of the code being added.

Patch-set: 12
Label: Code-Review=-1