Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > > Patch Set 12: Code-Review-1
> > > 
> > > I know you're working against a beta edition of the API, but I feel like we shouldn't be implementing the RTD access module this way. It would make sense for that to be it's own module on PyPi for others to be able to use as well!
> > > 
> > > Did you check to see if there an RTD module available already that we could use / perhaps work with to get the new API features built out in?
> > > 
> > > I see no unit tests in here for any of the code being added.
> > 
> > I agree going forward, but there are no existing implementations,we have somewhat of a time crunch with regards to implementing this for oran.
> 
> The urgency was mainly due to needing to do the other implementation bits, if you have time feel free to add some unit tests. I have a PR open for rtdv3 to implement a POST method for sub-projects that I'm hoping we can get it before the docs needed to be pushed.

My problem is that we've let this slide multiple times and hence a significant portion of lftools _has no tests_ We never have the time to go back and add them and then we make changes and it breaks things which is what unit tests are supposed to help prevent! The best time to add tests is at the point of absolutely _new_ implementation, not after the fact.

My review of -1 will continue to stand on that alone. I won't block this from going in, but I definitely won't give it a positive score with no tests in there. If you wan't to give it a merge, the code itself looks clean to me, but I just can't give it a positive vote knowing our history on this sort of thing.

Patch-set: 12