Update patch set 1
Patch Set 1: Code-Review-1

I think this is wrong. We _do_ need to exit 1 if we end up in one of the sections in which there is no DCO line where there should be. I think the problem is that the status=1 line is happening all the time if there are branches, and not IIF we find a condition in which it should be 1.

Patch-set: 1
Reviewer: Gerrit User 1 <1@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1