Update patch set 6
Patch Set 6:

(2 comments)

> Patch Set 6:
> 
> (2 comments)
> 
> whats the reason to do this with a yaml file? it seems like a one off operation I would prefer to pass all the needed on the command line. For automation creating a yaml file is cumbersome.
> Also could you provide a concrete example of how this would be used in the real world?

Creating from a file is part of the requirements for this feature: https://jira.linuxfoundation.org/browse/RELENG-290

Could you provide more details of what you'd like an example of? If there's something you're unsure of, it'd probably be best to add it to the docs, so that I'm answering that question for future users too.

Patch-set: 6
Label: Code-Review=0