Update patch set 6
Patch Set 6:

> > Patch Set 6: Code-Review+1
 > >
 > > Is this tested on any of the jobs? Do we have UT in place?
 > 
 > This is a refactor of other multithreaded upload code, tested
 > against a live nexus deployment.
 > 
 > UT?

UT(unit tests), I'd really like to see us python mock tests with the lftools, this helps with regressions. This is optional if you don't really have the cycles and something that can be done later.

I'll test this change with a job or two and then merge it over the weekend, it's not to hard to test an unmerged change. I've documented the steps below: :-) 

https://docs.releng.linuxfoundation.org/en/latest/jenkins.html#how-to-test-unmerged-cr-in-global-jjb-or-lftools-with-jenkins

Patch-set: 6