Update patch set 18
Patch Set 18: Code-Review-1

When testing this locally on a personal repo of mine, the match command fails to find any hashes when run as it currently is. Since check was working, I tried changing the get_branches call include "invert=True", and then it worked. I'll dive into what's happening in get_branches when it is not inverted, but for now this should not be merged.

Patch-set: 18
Reviewer: Gerrit User 192 <192@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=-1