Create gerrit project via gerrit api
Extend lfidapi to exit with good message and code if
group does not exist. (check if group exists functionality)

extend lftools gerrit to:
Create a project
Submit a .gitreview
Grant read to github user on repo
Create INFO.yaml patchset
TODO:
need to create github project, add replication user etc.
need to start replication

basically, if a new info file is submitted to info-master repo
git diff-tree --no-commit-id -r $rev --diff-filter=A --name-only --

Then we check if a gerrit repo exists
with checkproject

if not, we create it using the info file to set the owner.
then we can create the group with lfidapi
and invite or add base on provided info file

rinse and repeat for jira etc.

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I3a24751eeb2e739dee183074bbef6742df58a793