Create gerrit project via gerrit api
this works, but I want to force push a .gitreview, which I'm not sure
about via the api.
maybe I just submit it via the api. that seems doable.
I would also like a way to enable github replication.
but some file has to be submitted first.

basically, if a new info file is submitted to info-master repo
git diff-tree --no-commit-id -r $rev --diff-filter=A --name-only --

Then we check if a gerrit repo exists
with checkproject

if not, we create it using the info file to set the owner.
then we can create the group with lfidapi
and invite or add,

rinse and repeat for jira etc.

Very WIP

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I3a24751eeb2e739dee183074bbef6742df58a793