Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
>  > 
>  > Manually tested it and got the following result.
>  > $ lftools nexus release --verify-only --server https://nexus.opendaylight.org odlparent-2132
>  > Request URL: https://nexus.opendaylight.org/service/local/staging/repository/odlparent-2132/activity
>  > [<value>Missing: no javadoc jar found in folder '/org/opendaylight/odlparent/leveldbjni-all/6.0.1'</value>, <value>com.sonatype.nexus.staging.StagingRulesFailedException: One or more rules have failed</value>]
>  > One or more rules failed
>  > $
> 
> well this is a good start, however ... given the error this should print the rules that have failed. this might need something similar to using xmlstarlet programmatically.

The current code do not have anything pretty to print, apart from "Oups, failed" + the actual value that failed (StagingRulesFailedException or Invalid). I just added the "Missing" value to this chain.
The code is using "BeautifulSoup" to get values from an XML hierarchy.
Not quite sure what you would like to see.... Would you be able to elaborate?
I am not 100% certain just looking for the word "Missing:" is enough, same as looking for the word "Invalid". They might perhaps appear in a normal ok text message, but the code would detect this and flag failure.
I still ponder that the safest way would be to check for ruleFailed 
<stagingActivityEvent>
<timestamp>2019-10-18T09:39:24.841Z</timestamp>
<name>ruleFailed</name>
And in this structure pick up the failureMessage
<stagingProperty>
<name>failureMessage</name>
<value>
Missing: no javadoc jar found in folder '/org/opendaylight/odlparent/leveldbjni-all/6.0.1'
</value>
</stagingProperty>

Patch-set: 2
Label: Code-Review=0