Update patch set 1
Patch Set 1:

(2 comments)

I was blind for a moment there, sorry.

Please add a test for this specific condition - a link back to a parent directory that creates a loop.

Also I don't see a test to exercise the new code that tests if open can find a file, maybe you need to create a link to a non-existing file?

     ln -s /tmp/does-not-exist.log .

Patch-set: 1