Update patch set 3
Patch Set 3:

Sorry, I realized after pushing that last change that I don't need the extra dummy files in this test directory (they're there to make sure that only the correct artifacts are archived, but that's not what we're testing here). Removing them is the only change from patch 2.

Patch-set: 3